# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Site:
    properties:
      createdAt:
        type: string
      domain:
        type: string
      id:
        type: integer
      identifier:
        type: string
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.SiteMetaData'
      name:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SiteMetaData:
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.UserMetaData'
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      userType:
        $ref: '#/definitions/models.UserType'
    type: object
  models.UserMetaData:
    type: object
  models.UserType:
    enum:
    - ADMIN
    - PUBLISHER
    - ADVERTISER
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RolePublisher
    - RoleAdvertiser
  models.Zone:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      identifier:
        type: string
      metadata:
        $ref: '#/definitions/models.ZoneMetaData'
      name:
        type: string
      site:
        $ref: '#/definitions/models.Site'
      siteID:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
      zoneType:
        $ref: '#/definitions/models.ZoneType'
    type: object
  models.ZoneMetaData:
    type: object
  models.ZoneType:
    enum:
    - VIDEO
    - BANNER
    - NATIVE
    type: string
    x-enum-varnames:
    - ZoneTypeVideo
    - ZoneTypeBanner
    - ZoneTypeNative
host: localhost:8000
info:
  contact: {}
  description: REST API for AdAli.
  title: AdAli API
  version: "1.0"
paths:
...
  /zones:
    get:
      description: |-
        Get all zones or filter and order them using query parameters.

        Filtering operators:
        :   (equals)
        !:  (not equals)
        >   (greater than)
        >=  (greater than or equal)
        <   (less than)
        <=  (less than or equal)
        ~   (contains, case-insensitive)

        Examples:
        /zones?filter=is_active:true
        /zones?filter=site_id:1
        /zones?filter=name~banner
        /zones?order_by=created_at&sort=desc
        /zones?order_by=name&order_by=created_at&sort=asc&sort=desc
      parameters:
      - collectionFormat: csv
        description: Filter expression. Can be repeated.
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: Fields to order by. Can be repeated.
        in: query
        items:
          type: string
        name: order_by
        type: array
      - collectionFormat: csv
        description: Sort direction for each order_by field (asc or desc). Can be
          repeated.
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Zone'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get zones
      tags:
      - Zones
    post:
      consumes:
      - application/json
      description: Create a new zone.
      parameters:
      - description: Zone
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/models.Zone'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Zone'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create zone
      tags:
      - Zones
  /zones/{id}:
    delete:
      description: Delete a zone by ID.
      parameters:
      - description: Zone ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete zone
      tags:
      - Zones
    put:
      consumes:
      - application/json
      description: Update an existing zone.
      parameters:
      - description: Zone ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated zone
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/models.Zone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Zone'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update zone
      tags:
      - Zones
swagger: "2.0"
...